Fix the support of MacOS when we compile Bob natively
It seems that MacOS has the same behavior than FreeBSD. This patch
fix the pending TODO made when we homogenized the `connect()`
implementation. However, it seems that the behavior on Cosmopolitan was
not the same (with Cosmopolitan, on MacOS, the `connect()` works). It's
probably due to that `IsFreeBSD` probably return `true` on MacOS but
I'm not sure.