Introduce the --reproduce option
This option is only for debugging. Instead of using the volatile
identity of peers, we use "foo" and "bar" to be sure that the generation
of the shared-key is reproducible (as long as the --seed option is used
as well). However, for a proper usage, this option **is not** secure.