Improve the crypto module
This commit internalize ke to be able to logs our internal queue used by the
crypto module. It implements the CLOSE packet and the sender waits such
packet at the end of the transmission to be sure that the connection is
alive along the transmission.

For the second point, I suspect cloexec which sends FIN even if the
receiver did not transmit everything. The FIN is retransmitted to the
relay which abrubtely close the connection. The reason behind is because
we **can not** flush a TCP/IP socket and we must have a delimiter (like
QUIT with SMTP).

"\xff\xff\xff\xff" is the CLOSE packet. We are able to send only [<= 0x7fff]
packets now.