Add the new bob.happy-eyeballs package to handle correctly the `connect()` to a server.
Before this commit, if the DNS stack can not resolve a domain name, Bob
fails - even if the user specify multiple nameservers. With this commit,
Bob is able to try to resolve the domain name to **many** nameservers
(and not only one - by default, unicast.uncensoreddns.org). Now, the
behavior of Bob is more reliable than before when, if
unicast.uncensoreddns.org can not resolve osau.re, we still are able to
continue with 8.8.8.8 (for instance).