Return a new event: the identity given by the relay
To be able to use Tor, we must use something else to identify peers.
The identity is only specified by the relay now and peers must
acknowledge what the relay gives. A new event (as the Agreement) appears
mostly to notify up-front (on the CLI) the new identity of the peer via
our spelled word generator.