Handle the new identity event and show up into the CLI
This patch factorize also the way to process packets. Currently, a
TCP/IP packet can have multiple Bob's packets. We try to parse all of
them and aggregate events if we go them (`Identity and `Agreement).
According to our protocol, we can not have multiple `Identity or
multiple `Agreement and `Identity can not appear with `Agreement (in
this last case, we raise an error).

The idea is to show up the identity given by the relay to the sender and
the receiver can confirm the reception if the identity of the sender
matches with what the real sender gave to it.

We ensure that the relay can **not** generate the same identity for 2
active peers.